requirements.txt
Flask~=3.0.2
bcrypt~=4.1.2
pyJWT~=2.8.0
python-dotenv~=1.0.1
SQLAlchemy~=2.0.28
pydantic~=2.6.4
black
psycopg2
alembic
flask_cors
openai
Authlib
flask-socketio
stripe

